source code (1)/app_tier (1)/requirement (1).txt
boto3==1.21.13
botocore==1.24.13
certifi==2021.10.8
charset-normalizer==2.0.12
ec2-metadata==2.7.0
facenet-pytorch==2.5.2
idna==3.3
jmespath==0.10.0
numpy==1.22.0
Pillow==9.0.0
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
urllib3==1.26.8
